RASS-241 identified missing fields (#92)
* RASS-241 identified missing fields

* RASS-241 save missing fields

* update equals

* coverted period length to integer

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsremandandsentencingapi/controller/dto/CreatePeriodLength.kt
@@ -1,11 +1,9 @@
 package uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.controller.dto
 
-import java.math.BigDecimal
-
 data class CreatePeriodLength(
-  val years: BigDecimal?,
-  val months: BigDecimal?,
-  val weeks: BigDecimal?,
-  val days: BigDecimal?,
+  val years: Int?,
+  val months: Int?,
+  val weeks: Int?,
+  val days: Int?,
   val periodOrder: String,
 )
